refactor: change callAndRetry API to be like call
