+++ src/app/background/sagas/fetchRestrictedContexts.saga.ts
@@ -1,16 +1,13 @@
-import { put, call } from 'redux-saga/effects';
+import { put } from 'redux-saga/effects';
 import fetchRestrictedContexts from 'api/fetchRestrictedContexts';
 import { updateRestrictedContexts } from 'app/actions';
-import { regressiveRetry } from 'app/sagas/effects/regressiveRetry';
+import { createCallAndRetry } from 'app/sagas/effects/callAndRetry';
 
 export function* fetchRestrictedContextsSaga() {
-  const restrictedContexts = yield call(
-    regressiveRetry,
-    {
-      maximumRetryDelayInMinutes: 60
-    },
-    fetchRestrictedContexts
-  );
+  const callAndRetry = createCallAndRetry({
+    maximumRetryDelayInMinutes: 60
+  });
+  const restrictedContexts = yield callAndRetry(fetchRestrictedContexts);
 
   if (restrictedContexts) {
     yield put(updateRestrictedContexts(restrictedContexts));
