+++ src/app/background/sagas/subscriptions.ts
@@ -3,20 +3,17 @@ import { STARTUP, SUBSCRIBE, UNSUBSCRIBE } from 'app/actions';
 import { getSubscriptions } from 'app/background/selectors/subscriptions.selectors';
 import postSubscriptions from 'api/postSubscriptions';
 import { loginSaga } from './user.saga';
-import { regressiveRetry } from '../../sagas/effects/regressiveRetry';
+import { createCallAndRetry } from '../../sagas/effects/callAndRetry';
 
 function* postSubscriptionsSaga() {
   const extensionId = yield call(loginSaga);
   const subscriptions = yield select(getSubscriptions);
 
-  yield call(
-    regressiveRetry,
-    {
-      maximumAttempts: 10
-    },
-    postSubscriptions,
-    { extensionId, subscriptions }
-  );
+  const callAndRetry = createCallAndRetry({
+    maximumAttempts: 10
+  });
+
+  yield callAndRetry(postSubscriptions, { extensionId, subscriptions });
 }
 
 export default function*() {
